Integrated multiple strips code with the local prediction.
This involves three main changes.
(1)The async task launching the connections to the servers is exchanged
with a task that predicts.
(2) A hash key is computed on the location to asure that is correctly
represented in the library.
(3) The timers for the sendings and the receives are no longer needed
thus eliminated.